+++ Source/ACE.AuthApi/Controllers/AccountController.cs
@@ -12,6 +12,9 @@ using System.Web.Http;
 
 namespace ACE.AuthApi.Controllers
 {
+    /// <summary>
+    ///
+    /// </summary>
     public class AccountController : BaseController
     {
         /// <summary>
@@ -45,6 +48,9 @@ namespace ACE.AuthApi.Controllers
             return account;
         }
 
+        /// <summary>
+        ///
+        /// </summary>
         [HttpPost]
         [AllowAnonymous]
         [SwaggerResponse(HttpStatusCode.OK, "Auth successful", typeof(RegisterResponse))]
@@ -53,6 +59,9 @@ namespace ACE.AuthApi.Controllers
             return Request.CreateResponse(HttpStatusCode.NotImplemented);
         }
 
+        /// <summary>
+        ///
+        /// </summary>
         [HttpGet]
         [AceAuthorize]
         [SwaggerResponse(HttpStatusCode.OK, Type = typeof(Account))]