+++ Source/ACE.AuthApi/Controllers/RegisterResponse.cs
@@ -7,17 +7,32 @@ using System.Threading.Tasks;
 
 namespace ACE.AuthApi.Controllers
 {
+    /// <summary>
+    ///
+    /// </summary>
     public class RegisterResponse
     {
+        /// <summary>
+        ///
+        /// </summary>
         [JsonProperty("username")]
         public string Username { get; set; }
 
+        /// <summary>
+        ///
+        /// </summary>
         [JsonProperty("authToken")]
         public string AuthToken { get; set; }
 
+        /// <summary>
+        ///
+        /// </summary>
         [JsonProperty("accountId")]
         public uint AccountId { get; set; }
 
+        /// <summary>
+        ///
+        /// </summary>
         [JsonProperty("accountGuid")]
         public Guid AccountGuid { get; set; }
     }