+++ Source/ACE.AuthApi/SwaggerConfig.cs
@@ -9,8 +9,14 @@ using ACE.AuthApi;
 
 namespace ACE.AuthApi
 {
+    /// <summary>
+    ///
+    /// </summary>
     public class SwaggerConfig
     {
+        /// <summary>
+        ///
+        /// </summary>
         [SuppressMessage("Microsoft.StyleCop.CSharp.Spacing", "SA1005:SingleLineCommentsMustBeginWithSingleSpace")]
         public static void Register(HttpConfiguration config)
         {