+++ Source/ACE.Diagnostics/Program.cs
@@ -5,13 +5,13 @@ using System.Windows.Forms;
 
 namespace ACE.Diagnostics
 {
-    static class Program
+    public static class Program
     {
         /// <summary>
         /// The main entry point for the application.
         /// </summary>
         [STAThread]
-        static void Main()
+        public static void Main()
         {
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);