+++ Source/ACE.Entity/Enum/Properties/PropertyString.cs
@@ -78,5 +78,3 @@ namespace ACE.Entity.Enum.Properties
         }
     }
 }
-
-